v0.12.1
Fixing small issues refactoring js files, fixing migrations

+++ db/schema.rb
@@ -18,7 +18,7 @@ ActiveRecord::Schema.define(version: 2021_02_15_094002) do
   create_table "answers", force: :cascade do |t|
     t.string "title", null: false
     t.boolean "correct", default: false
-    t.integer "question_id", null: false
+    t.bigint "question_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["question_id"], name: "index_answers_on_question_id"
@@ -42,7 +42,7 @@ ActiveRecord::Schema.define(version: 2021_02_15_094002) do
 
   create_table "questions", force: :cascade do |t|
     t.string "title", null: false
-    t.integer "test_id", null: false
+    t.bigint "test_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["test_id"], name: "index_questions_on_test_id"
@@ -63,8 +63,8 @@ ActiveRecord::Schema.define(version: 2021_02_15_094002) do
   create_table "tests", force: :cascade do |t|
     t.string "title", null: false
     t.integer "level", default: 0
-    t.integer "category_id", null: false
-    t.integer "user_id", null: false
+    t.bigint "category_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["category_id"], name: "index_tests_on_category_id"
